Refactor migration files to include checks for existing columns and t…

+++ database/migrations/2025_03_21_061218_student_answers.php
@@ -12,7 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         //
-        Schema::create('student_answers', function (Blueprint $table) {
+        if (!Schema::hasTable('student_answers')) {
+            Schema::create('student_answers', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('task_id')->constrained()->onDelete('cascade');
@@ -20,6 +21,7 @@ return new class extends Migration
             $table->boolean('is_correct')->default(false);
             $table->timestamps();
         });
+        }
     }
 
     /**
